Some fixes for GenericDatum type: 1. Allow more than 2 types in a union. 2. Fix Validate for RecordSchema 3. Fix string length encoding: int32-int64 mismatch 4. Fix enum parsing for GenericDatum.

+++ decoder.go
@@ -154,14 +154,14 @@ func (this *BinaryDecoder) ReadString() (string, error) {
 	if err := checkEOF(this.buf, this.pos, 1); err != nil {
 		return "", err
 	}
-	length, err := this.ReadInt()
+	length, err := this.ReadLong()
 	if err != nil || length < 0 {
 		return "", InvalidStringLength
 	}
 	if err := checkEOF(this.buf, this.pos, int(length)); err != nil {
 		return "", err
 	}
-	value := string(this.buf[this.pos : int32(this.pos)+length])
+	value := string(this.buf[this.pos : this.pos+length])
 	this.pos += int64(length)
 	return value, nil
 }
